cleaned up file paths for correlation files, tested script works
